feat: First changes for a clear simple agent for workshop
